% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_vector_direction.R
\name{calculate_vector_direction}
\alias{calculate_vector_direction}
\title{Front wave velocity direction}
\usage{
calculate_vector_direction(x, y)
}
\arguments{
\item{x}{The magnitude of a vector in x}

\item{y}{The magnitude of a vector in y}
}
\description{
Determines the direction of a vector given magnitude in x and y direction
}

