% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_velocity.R
\name{calculate_velocity}
\alias{calculate_velocity}
\title{Calculate the front-wave velocity}
\usage{
calculate_velocity(d)
}
\arguments{
\item{d}{The outbreak rate functions returned from outbreak_rate_functions}
}
\description{
Calculates the front-wave velocity of an outbreak given the outbreak rate functions
}

