% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_surfacetrend_models.R
\name{estimate_surfacetrend_models}
\alias{estimate_surfacetrend_models}
\title{Fit polynomial models}
\usage{
estimate_surfacetrend_models(ds, max.order = 10)
}
\arguments{
\item{ds}{Dataframe providing the date of outbreak and X and Y coordinates}

\item{max.order}{Integer of highest order polynomial to attempt; defaults to 10}
}
\description{
This function fits the polynomia functions of the surface trend analysis.
}

