% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_model.R
\name{get_best_model}
\alias{get_best_model}
\title{Get the best fitting surface trend model}
\usage{
get_best_model(summary, measure = "r2")
}
\arguments{
\item{summary}{Dataframe returned from function surfacetrend_model_fit}

\item{measure}{The measure of model fit, defaults to R-squared ('r2'), can be 'aic' or 'bic'}
}
\description{
Retrieves the order of the best fitting mdoel based on the given measure of model fit
}

