% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_rate_functions.R
\name{outbreak_rate_functions}
\alias{outbreak_rate_functions}
\title{Take the derivatives of the polynomial model}
\usage{
outbreak_rate_functions(m, measure = "r2", x, y, order)
}
\arguments{
\item{m}{An lm model object}

\item{measure}{The measure of model fit, defaults to R-squared, 'r2', can be 'aic' or 'bic'}

\item{x}{Units indicating the position in x-axis of the outbreak points}

\item{y}{Units indicating the position in y-axis of the outbreak points}

\item{order}{The order of the polynomial selected for the surface trend}

\item{y}{Units indicating the position in y-axis of the outbreak points}
}
\description{
Gets the functions of the partial derivatives with respect to X and Y to estimate the front-wave velocity of the outbreak
}

