% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_velocity.R
\name{outbreak_velocity}
\alias{outbreak_velocity}
\title{The front-wave velocity}
\usage{
outbreak_velocity(ds, max.order = 10, measure = "r2",
  manual.order = FALSE)
}
\arguments{
\item{ds}{Dataframe providing the date of outbreak and X and Y coordinates}

\item{max.order}{Integer of highest order polynomial to attempt; defaults to 10}

\item{measure}{The measure of model fit, defaults to R-squared: 'r2', can be AIC: 'aic' or BIC: 'bic'}

\item{manual.order}{If TRUE, the max.order integer is the polynomial model used, regardless of best fit measures; defaults to FALSE}
}
\description{
Calculates the front-wave velocity of an outbreak
}

