% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfacetrend_model_fit.R
\name{surfacetrend_model_fit}
\alias{surfacetrend_model_fit}
\title{Assess the model fit}
\usage{
surfacetrend_model_fit(lm.fit)
}
\arguments{
\item{lm.fit}{A list of lm model objects}
}
\description{
This function assess the fit of the surface trend polynomial regression models by AIC, BIC, and R-squared
}

