% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_vector_map.R
\name{velocity_vector_map}
\alias{velocity_vector_map}
\title{The front-wave velocity vector map}
\usage{
velocity_vector_map(r, shpfile, fid, dsid)
}
\arguments{
\item{r}{The front-wave velocity summary from the \code{\link{outbreak_velocity}}  function}

\item{shpfile}{A polynomial shapefile object \code{"SpatialPolygonsDataFrame"} from maptools}

\item{fid}{The name (string) of a unique numeric ID for each field in the shpfile; must correspond to \code{"id"} and must have one element for every value of \code{"id"}}

\item{dsid}{A vector of the unique numeric ID for each field in the original dataset; must correspond to \code{"fid"}}
}
\description{
Plots the magnitude and direction of the outbreak front-wave velocity
}

